Fix for Soaring and Walking Mode
This adds additional checks to the walking mode in the Gen 6 SR bot to
make sure that the step toward the Mirage Spot or Legendary is taken,
making the bot much more stable.
It also makes the Soaring bot much more reliable because it's now not
going to stop if it gets into a random encounter in the sky anymore.