Added easy modification of the enableillegal bool.
Added the ability for the end user to easily enable the writing of illegal Pokemon.

Change the name of PKMN-NTR.exe to PKMN-NTR-ILLEGAL.exe to enabled writing of illegal Pokemon. Name the executable to anything else to disable the writing of illegal Pokemon.

You'll be notified on launch if you correctly enabled the writing of illegal Pokemon.